Adding submodule as a pytest-plugin with fixtures and assert helpers
