+++ packages/service-integration/requirements/_base.in
@@ -2,24 +2,16 @@
 # Specifies third-party dependencies for 'service-integration'
 #
 
-
+# tools
+pyyaml
 click
 
 # publising
 #bump2version
-
-
-
-# tools
-pyyaml
-#docker
 #jsonschema
 
 
-# tests
-#pytest
-#pytest-cookies
-#pytest-cov
-#pytest-instafail
-#pytest-mock
-#pytest-sugar
+# pytest-plugin
+pytest
+docker
+jsonschema
